+++ webpack.config.js
@@ -4,30 +4,42 @@ const MiniCssExtractPlugin = require('mini-css-extract-plugin');
 const CopyPlugin = require('copy-webpack-plugin');
 const HtmlConfig = require(path.join(__dirname, 'html.config'));
 const CleanPlugin = require('clean-webpack-plugin');
+const ESLintPlugin = require('eslint-webpack-plugin');
 
 const deploy = path.join(__dirname, 'deploy');
+const isProduction = process.env.NODE_ENV == "production";
 
-module.exports = env => {
+// keep the env param to be explicit, eslint disable should be removed when template is in use
+// eslint-disable-next-line no-unused-vars
+module.exports = (env) => {
   const plugins = [
     new CleanPlugin.CleanWebpackPlugin(),
     new HtmlWebpackPlugin(HtmlConfig),
     new MiniCssExtractPlugin({ filename: 'css/game.style.css' }),
-    new CopyPlugin([{ from: path.join(__dirname + '/static') }])
+    new CopyPlugin({
+      patterns: [
+        { from: path.join(__dirname + '/static'), to: deploy }
+      ]
+    }),
+    new ESLintPlugin()
   ];
+
   return {
     stats: 'errors-only',
 
-    mode: env.dev ? 'development' : 'production',
+    mode: isProduction ? 'production':'development',
 
     devServer: {
       open: true,
-      overlay: true,
+      client: { overlay: true },
       host: '0.0.0.0',
-      public: 'localhost:8080',
-      contentBase: path.join(__dirname, '/static')
+      port: 8080,
+      static: path.join(__dirname, '/static')
     },
 
-    entry: ['@babel/polyfill', path.join(__dirname, '/src/index.js')],
+    context: path.resolve(__dirname, 'src'),
+
+    entry: path.join(__dirname, '/src/index.js'),
     output: {
       filename: 'js/game.bundle.js',
       path: deploy
@@ -54,7 +66,7 @@ module.exports = env => {
         {
           test: /\.js$/,
           exclude: /(node_modules|bower_components)/,
-          use: ['babel-loader', 'eslint-loader']
+          use: ['babel-loader']
         },
         {
           test: /\.(otf|woff(2)?|ttf|eot|svg)(\?v=\d+\.\d+\.\d+)?$/,
@@ -100,7 +112,12 @@ module.exports = env => {
               }
             }
           ]
-        }
+        },
+        {
+          test: /\.js$/,
+          enforce: 'pre',
+          use: ['source-map-loader'],
+        },
       ]
     }
   };