+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2022-12-12
+
+### Changed
+
+- Updated webpack to 5.75.0
+- Updated webpack packages to match v5
+- Bump loader-utils from 2.0.3 to 2.0.4
+
+### Added
+
+- Added eslint rules
+
 ## [1.1.2] - 2022-12-02
 
 ### Changed